refactor: extract shared run_vpp, environment inputs, and validation
Move run_vpp, render_environment_inputs, and validate_ranges into
demos/utils.py so both the VPP and Compare pages use the same code.
Removes ~40 lines of duplication across the two pages.
